+++ packages/contracts/src/support/index.ts
@@ -4,3 +4,5 @@
  * @type {Symbol}
  */
 export const SUPPORT: unique symbol = Symbol('@aedart/contracts/support');
+
+export type * from './types';