+++ packages/contracts/src/support/types.ts
@@ -1,0 +1,14 @@
+/**
+ * Key or path identifier
+ */
+export type Key = OneOrMany<PropertyKey>;
+
+/**
+ * Object Property key
+ */
+export type PropertyKey = string | number | symbol;
+
+/**
+ * One or many
+ */
+export type OneOrMany<T> = T | ReadonlyArray<T>;